+++ tests/config_template.py
@@ -2,7 +2,7 @@
 !!! COPY THIS FILE TO config.py !!!
 
 Replace value of `directory` with the path to a folder containing .xer files.
-The test setup will recursively run through this folder and any children folders to find
+The test setup will recursively run through this folder and any subfolders to find
 all .xer files.
 
 Caution: the more .xer files found, the longer the setup and test will take to run.